refactor: Line 생성하는 과정을 LadderCreator로 분리
