+++ src/main/java/ladder/domain/Ladder.java
@@ -1,20 +1,15 @@
 package ladder.domain;
 
-public class Ladder {
-    private final int countOfPerson;
-    private final int ladderHeight;
+import java.util.List;
 
-    public Ladder(int countOfPerson,int ladderHeight) {
-        this.countOfPerson = countOfPerson;
-        this.ladderHeight = ladderHeight;
+public class Ladder {
+    private final List<Line> lineList;
+
+    public Ladder(final List<Line> lineList) {
+        this.lineList = lineList;
     }
 
-    public LineList createLadder() {
-        LineList lineList = new LineList();
-        for (int idx = 0; idx < ladderHeight; idx++) {
-            lineList.createLine(countOfPerson);
-        }
+    public List<Line> getLineList() {
         return lineList;
     }
-
 }
